Boi
Defined the MD5 and most Placeholder fonts to the purple color/mint color. Not sure if you wanted me to, but I did it anyway. Used the mint to defer which ones are which, but feel free to change the class defined color back to purple if you wish. I didnt do all of the lines becasue im unsure if it was okay to change the fonts you had set.